+++ lib/ts/recipe/thirdpartyemailpassword/types.ts
@@ -36,8 +36,8 @@ import {
 } from "../thirdparty";
 import {
     NormalisedConfig as TPConfig,
-    StateObject,
     PreAndPostAPIHookContext as ThirdPartyPreAndPostAPIHookAction,
+    StateObject,
 } from "../thirdparty/types";
 import Provider from "../thirdparty/providers";
 import { CustomProviderConfig } from "../thirdparty/providers/types";
@@ -139,8 +139,6 @@ export type SignInAndUpInput =
       }
     | {
           type: "thirdparty";
-          thirdPartyId: string;
-          authCode?: string;
           config: TPConfig;
           userContext: any;
       };
@@ -179,7 +177,6 @@ export type RecipeInterface = {
             id: string;
             value: string;
         }[];
-        token: string;
         config: EPConfig;
         userContext: any;
     }) => Promise<
@@ -225,11 +222,9 @@ export type RecipeInterface = {
         fetchResponse: Response;
     }>;
 
-    getOAuthAuthorisationURLFromBackend: (input: {
-        thirdPartyId: string;
-        config: TPConfig;
-        userContext: any;
-    }) => Promise<{
+    getSubmitPasswordTokenFromURL: (input: { config: EPConfig; userContext: any }) => string;
+
+    getAuthorisationURLFromBackend: (input: { thirdPartyId: string; config: TPConfig; userContext: any }) => Promise<{
         status: "OK";
         url: string;
         fetchResponse: Response;
@@ -237,30 +232,26 @@ export type RecipeInterface = {
 
     signInAndUp: (input: SignInAndUpInput) => Promise<SignInAndUpOutput>;
 
-    getStateAndOtherInfoFromStorage: <CustomStateProperties>(input: {
-        userContext: any;
-        config: TPConfig;
-    }) => (StateObject & CustomStateProperties) | undefined;
+    getStateAndOtherInfoFromStorage: (input: { userContext: any; config: TPConfig }) => StateObject | undefined;
 
-    setStateAndOtherInfoToStorage: <CustomStateProperties>(input: {
-        state: StateObject & CustomStateProperties;
-        config: TPConfig;
-        userContext: any;
-    }) => void;
+    setStateAndOtherInfoToStorage: (input: { state: StateObject; config: TPConfig; userContext: any }) => void;
 
     redirectToThirdPartyLogin: (input: {
         thirdPartyId: string;
         config: TPConfig;
-        state?: StateObject;
         userContext: any;
     }) => Promise<{ status: "OK" | "ERROR" }>;
 
     generateStateToSendToOAuthProvider: (input: { userContext: any; config: TPConfig }) => string;
 
-    verifyStateFromOAuthProvider: (input: {
-        stateFromProvider: string | undefined;
-        stateFromStorage: string | undefined;
+    verifyAndGetStateOrThrowError: (input: {
+        stateFromAuthProvider: string | undefined;
+        stateObjectFromStorage: StateObject | undefined;
         config: TPConfig;
         userContext: any;
-    }) => boolean;
+    }) => Promise<StateObject>;
+
+    getAuthCodeFromURL: (input: { config: TPConfig; userContext: any }) => string;
+
+    getAuthErrorFromURL: (input: { config: TPConfig; userContext: any }) => string | undefined;
 };