+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Exports more recipe functions for emailverification recipe to allow them to be called without using the pre-built UI. Newly exported functions: `verifyEmail`, `sendVerificationEmail`
 -   Exports all emailverification recipe functions from emailpassword, thirdparty and thirdpartyemailpassword recipes.
 -   Exports more recipe functions for emailpassword recipe to allow them to be called without using the pre-built UI. Newly exported functions: `submitNewPassword`, `sendPasswordResetEmail`, `signUp`, `signIn`, `doesEmailExist`.
+-   Exports more recipe functions for thirdparty recipe to allow them to be called without using the pre-built UI Newly exported function: `redirectToThirdPartyLogin`, `signInAndUp`.
 
 ### Breaking changes